Rename PFParticle to PFTruthParticle
