+++ SimDataFormats/PFAnalysis/src/classes.h
@@ -2,7 +2,7 @@
 #include "SimDataFormats/TrackingAnalysis/interface/TrackingParticleFwd.h"
 #include "SimDataFormats/CaloAnalysis/interface/SimCluster.h"
 #include "SimDataFormats/CaloAnalysis/interface/SimClusterFwd.h"
-#include "SimDataFormats/PFAnalysis/interface/PFParticle.h"
-#include "SimDataFormats/PFAnalysis/interface/PFParticleFwd.h"
+#include "SimDataFormats/PFAnalysis/interface/PFTruthParticle.h"
+#include "SimDataFormats/PFAnalysis/interface/PFTruthParticleFwd.h"
 #include "DataFormats/Common/interface/Wrapper.h"
