+++ SimDataFormats/PFAnalysis/src/classes_def.xml
@@ -1,12 +1,12 @@
 <lcgdict>
-  <class name="PFParticle" />
-  <class name="std::vector<PFParticle>" />
-  <class name="edm::Wrapper<std::vector<PFParticle> >" />
-  <class name="edm::RefProd<std::vector<PFParticle> >" />
-  <class name="PFParticleRef" />
-  <class name="edm::Ptr<PFParticle>" />
-  <class name="std::vector<edm::Ptr<PFParticle>>" />
-  <class name="PFParticleRefVector" />
-  <class name="edm::Wrapper<PFParticleRefVector>"/>
+  <class name="PFTruthParticle" />
+  <class name="std::vector<PFTruthParticle>" />
+  <class name="edm::Wrapper<std::vector<PFTruthParticle> >" />
+  <class name="edm::RefProd<std::vector<PFTruthParticle> >" />
+  <class name="PFTruthParticleRef" />
+  <class name="edm::Ptr<PFTruthParticle>" />
+  <class name="std::vector<edm::Ptr<PFTruthParticle>>" />
+  <class name="PFTruthParticleRefVector" />
+  <class name="edm::Wrapper<PFTruthParticleRefVector>"/>
 </lcgdict>
